# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vinyl Radar API",
    "description": "국내 바이닐 발매·재고 정보 통합 API",
    "version": "0.1.0"
  },
  "paths": {
    "/admin/releases": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Releases",
        "description": "초안을 포함한 전체 일정.",
        "operationId": "list_releases_admin_releases_get",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReleaseAdminOut"
                  },
                  "title": "Response List Releases Admin Releases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Release",
        "description": "일정을 등록한다. **기본은 초안**이며 공개하려면 별도 호출이 필요하다.",
        "operationId": "create_release_admin_releases_post",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/releases/{release_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Release",
        "description": "일정 하나. 초안이어도 운영자에게는 보인다 (공개 API 와 다른 점).",
        "operationId": "get_release_admin_releases__release_id__get",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Release",
        "operationId": "update_release_admin_releases__release_id__patch",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Release",
        "description": "**초안만** 삭제할 수 있다. 공개 중이면 먼저 공개를 취소해야 한다 (T-133).\n\n이전에는 \"한 번도 공개된 적 없는\" 것만 지울 수 있었다. 발송 이력을 지키려는\n규칙이었지만, 그 결과 **운영자가 남은 방법이 SQL 뿐**이 되었다.\n손으로 여러 테이블을 순서대로 지우는 쪽이 훨씬 위험하다 — 한 줄 틀리면\n남의 데이터까지 지운다. 그래서 가드를 여기로 옮긴다.\n\n**공개 중인 것은 여전히 못 지운다.** 공개 취소를 한 번 거치게 해서,\n실수로 살아 있는 일정을 지우는 일이 없도록 한 단계를 남긴다.\n\n감수하는 것 — 이 발매에 대해 **무엇을 언제 보냈는지 기록이 사라진다.**\n`notification_deliveries` 는 `listing_events` 의 CASCADE 로 함께 지워진다.",
        "operationId": "delete_release_admin_releases__release_id__delete",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/releases/{release_id}/publish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Publish Release",
        "description": "공개한다. 이 시점에 `SCHEDULE_ADDED` 이벤트가 **한 번만** 생긴다.",
        "operationId": "publish_release_admin_releases__release_id__publish_post",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/releases/{release_id}/unpublish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Unpublish Release",
        "description": "공개를 취소해 초안으로 되돌린다.\n\n잘못 공개한 일정을 되돌릴 수단이 없으면 삭제도 못 한다\n(공개본 삭제는 409 로 막혀 있다).\n\n**이미 나간 `SCHEDULE_ADDED` 이벤트는 지우지 않는다.** 구독자에게 전달된 사실은\n남아야 하고, 다시 공개해도 이벤트가 또 생기지는 않는다.",
        "operationId": "unpublish_release_admin_releases__release_id__unpublish_post",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseAdminOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/releases/{release_id}/links": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add Link",
        "operationId": "add_link_admin_releases__release_id__links_post",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseLinkIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseLinkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/releases/{release_id}/links/{link_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Link",
        "description": "구매처 링크를 지운다. 여러 개를 등록할 수 있으니 지우는 수단도 있어야 한다.",
        "operationId": "delete_link_admin_releases__release_id__links__link_id__delete",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Link Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/releases": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "List Releases",
        "description": "공개된 발매 일정 목록.\n\n정렬은 **예약 시작이 임박한 순**이며, 날짜 미정 일정은 맨 뒤로 간다.",
        "operationId": "list_releases_v1_releases_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "이전 응답의 next_cursor",
              "title": "Cursor"
            },
            "description": "이전 응답의 next_cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          },
          {
            "name": "is_limited",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Limited"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleasePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/releases/{release_id}": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Get Release",
        "description": "공개된 일정 하나.\n\n초안은 **404** 로 응답한다 — 403 으로 구분하면 \"존재는 한다\"는 사실이 새어 나간다.",
        "operationId": "get_release_v1_releases__release_id__get",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReleaseLinkIn": {
        "properties": {
          "shop_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Shop Name"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "price_krw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Krw"
          }
        },
        "type": "object",
        "required": [
          "shop_name",
          "url"
        ],
        "title": "ReleaseLinkIn",
        "description": "구매 링크 입력."
      },
...
    }
  }
}